backend/requirements.txt
fastapi==0.119.1
importlib-metadata==8.0.0
jaraco.collections==5.1.0
langchain==1.0.2
langchain-openai==1.0.1
pgvector==0.4.1
pip-chill==1.0.3
platformdirs==4.2.2
psycopg2-binary==2.9.11
pydantic-settings==2.11.0
sqlalchemy==2.0.44
tomli==2.0.1
uvicorn==0.38.0
